requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pydantic==2.5.0
pydantic[email]==2.5.0
bcrypt==4.1.1
python-dotenv==1.0.0
PyJWT==2.8.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
pytest
pytest-asyncio
httpx
mongomock
faker